[TL42-112] fix: 채팅방 컴포넌트 분리
 - '페이지'만 Chat에 남겨두고, ChatModal은 Templates로 분리.
