+++ front-end/src/UI/Pages/Chat/index.tsx
@@ -1,19 +1,7 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import {
-  Button,
-  Input,
-  InputGroup,
-  InputRightElement,
-  ModalBody,
-  ModalFooter,
-  ModalHeader,
-} from '@chakra-ui/react';
-import { ArrowBackIcon } from '@chakra-ui/icons';
-import RoutedModal from '../../Templates/RoutedModal';
-import ChatHeader from '../../Organisms/ChatHeader';
-import ChatBody from '../../Organisms/ChatBody';
 import { ChatProvider, useChat } from '../../../Hooks/useChat';
+import ChatModal from '../../Templates/ChatModal';
 
 function ChatExample() {
   const { id } = useParams();
@@ -183,28 +171,11 @@ function ChatExample() {
   return <div />;
 }
 
-export default function ChatModal() {
+export default function Chat() {
   return (
     <ChatProvider>
       <ChatExample />
-      <RoutedModal closeOnOverlayClick={false}>
-        <ModalHeader display={{ base: 'none', lg: 'flex' }}>
-          <ChatHeader />
-        </ModalHeader>
-        <ModalBody pb={6}>
-          <ChatBody />
-        </ModalBody>
-        <ModalFooter>
-          <InputGroup size="lg">
-            <Input pr="4.5rem" textColor="black" placeholder="" />
-            <InputRightElement width="3rem">
-              <Button colorScheme="gray" size="1.5em" p={1}>
-                <ArrowBackIcon boxSize="1.5em" />
-              </Button>
-            </InputRightElement>
-          </InputGroup>
-        </ModalFooter>
-      </RoutedModal>
+      <ChatModal />
     </ChatProvider>
   );
 }
