+++ front-end/src/UI/Pages/Main/index.tsx
@@ -6,7 +6,7 @@ import MainStandby from '../../Templates/MainStandby';
 import { SocketState, useSocket } from '../../../Hooks/useSocket';
 import Loading from '../../Templates/Loading';
 import RoutedModalExample from '../RoutedModalExample';
-import ChatModal from '../ChatModal';
+import Chat from '../Chat';
 
 function Main() {
   const { state } = useSocket();
@@ -43,7 +43,7 @@ function Main() {
       </Box>
       <Routes>
         <Route path="/example/:name" element={<RoutedModalExample />} />
-        <Route path="/chat/:id" element={<ChatModal />} />
+        <Route path="/chat/:id" element={<Chat />} />
       </Routes>
     </Flex>
   );
